Add atomic wrappers for Bool and Pointer, fix UInt64 String method
Introduced `Bool` and `Pointer` types with thread-safe operations, including `Get`, `Set`, `CompareAndSwap`, and conditional setters. Updated `UInt64` `String` method to simplify conversion logic.

+++ notatomic/atomicInt.go
@@ -396,5 +396,5 @@ func (a *UInt64) Toggle(mask uint64) uint64 {
 }
 
 func (a *UInt64) String() string {
-	return strconv.FormatUint(uint64(a.Get()), 10)
+	return strconv.FormatUint(a.Get(), 10)
 }
